Parallel bootstrap
Added ncpus (number of CPUs) to boot function. Now can specify how many
cpus to use. Use multicore for unix platform and snow for windows
platform. Still need to work out clustering for Windows as functions
that are called in CR.bs (e.g. CR_RobustDesign) are not initialised in
the separate workspaces that are created by snow for parallel processes
- causes errors for ncpus>1.